move fields and methods to PartyPokemon from ActivePokemon
